ClickTester
Execute the app a long or short press on any of the four colored button
in layout it should produce an appropriate Toast display indicating the
nature of the action. Example: A short finger tap on the yellow block at
the lower left has just been completed and it display that message on
the label at the bottom.